Create Introduction
#INTRODUCTION

According to the World Health Organization approximately 1.35 fatalities occur due to road accidents and around 20-50 million suffer injuries. Most of these mishaps are avoidable by taking stringent measures with the backing of through research.
Along with fatalities and injuries, road accidents also cost nations economically too in the form of damages and traffic delays. Hence, any nation would only benefit by improving the traffic safety and research.
This study will be based in the city of Seattle, Washington. Being one of the busiest cities in Westcoast, it sees a very high frequency of car accidents.

##BUSINESS PROBLEM

The main aim of this study is to be able to predict the severity of road accidents in Seattle , Washigton. The predictions will be based on road accident data collected by the WSDOT(Washington State Department of Transportation) which has archived data of road mishaps from the year 2004 and includes data regarding the weather conditions, traffic events, time etc.
The study will make use of several analytical models and techniques to predict the severity of road accidents.  

this study will help the WSDOT to take informed measures and make iprovements to the current traffic policies.